requirements.txt
flask
flask-bcrypt
supabase
python-dotenv
requests
python-dateutil
pycryptodome